represent array type, index and item as a map in json error results
